refactor(policies): extract AuthImage and normalizeSlateHtml from PolicyEditorPage
Move the TipTap auth image node view/extension and legacy Slate HTML normalizer into co-located PolicyEditor modules to start shrinking the page.
